+++ app/src/main/java/softeng/eventplanning/MainActivity.java
@@ -62,38 +62,6 @@ public class MainActivity extends AppCompatActivity {
 
     }
 
-    @Override
-    public boolean onCreateOptionsMenu(Menu menu) {
-        // Inflate the menu; this adds items to the action bar if it is present.
-        getMenuInflater().inflate(R.menu.settings_menu, menu);
-        return true;
-    }
-
-    @Override
-    public boolean onOptionsItemSelected(MenuItem item) {
-        // Handle action bar item clicks here. The action bar will
-        // automatically handle clicks on the Home/Up button, so long
-        // as you specify a parent activity in AndroidManifest.xml.
-        int id = item.getItemId();
-
-        //noinspection SimplifiableIfStatement
-        switch (id) {
-            case R.id.settings_notifications:
-                // TODO go to notifications page
-                Toast.makeText(getApplicationContext(),
-                        "Settings Updated",
-                        Toast.LENGTH_SHORT).show();
-                break;
-            case R.id.settings_themes:
-                // TODO go to theme page
-                Toast.makeText(getApplicationContext(),
-                        "Settings Updated",
-                        Toast.LENGTH_SHORT).show();
-                break;
-        }
-        return super.onOptionsItemSelected(item);
-    }
-
     public void eventClicked(View view){
         // TODO open event's page when clicked
     }
