+++ app/src/main/java/softeng/eventplanning/tabView.java
@@ -3,11 +3,14 @@ package softeng.eventplanning;
 import android.content.Intent;
 import android.os.Bundle;
 import android.support.v7.app.AppCompatActivity;
+import android.view.Menu;
+import android.view.MenuItem;
 import android.view.View;
 import android.widget.ArrayAdapter;
 import android.widget.ExpandableListView;
 import android.widget.ListView;
 import android.widget.TabHost;
+import android.widget.Toast;
 
 import java.util.ArrayList;
 import java.util.HashMap;
@@ -65,6 +68,37 @@ public class tabView extends AppCompatActivity {
         Intent intent = new Intent(this, sFilters.class);
         startActivity(intent);
     }
+    /*@Override
+    public boolean onCreateOptionsMenu(Menu menu) {
+        // Inflate the menu; this adds items to the action bar if it is present.
+        getMenuInflater().inflate(R.menu.settings_menu, menu);
+        return true;
+    }
 
+    @Override
+    public boolean onOptionsItemSelected(MenuItem item) {
+        // Handle action bar item clicks here. The action bar will
+        // automatically handle clicks on the Home/Up button, so long
+        // as you specify a parent activity in AndroidManifest.xml.
+        int id = item.getItemId();
+
+        //noinspection SimplifiableIfStatement
+        switch (id) {
+            case R.id.settings_notifications:
+                // TODO go to notifications page
+                Toast.makeText(getApplicationContext(),
+                        "Settings Updated",
+                        Toast.LENGTH_SHORT).show();
+                break;
+            case R.id.settings_themes:
+                // TODO go to theme page
+                Toast.makeText(getApplicationContext(),
+                        "Settings Updated",
+                        Toast.LENGTH_SHORT).show();
+                break;
+        }
+        return super.onOptionsItemSelected(item);
+    }
+*/
 
 }
